+++ CHANGELOG.md
@@ -400,3 +400,4 @@
 * Move `Vertical` datatype from `Binary` to `Morphable` module
 * Remove `substitute`, `subplace`, `subview` methods from `Substructure` typeclass
 * Remove `>>=:>` and `<:=<<` infix operators
+* Move `Vector` datatype to `Linear` umbrella-module